+++ crates/module-system/sov-modules-api/src/gas/tests/mod.rs
@@ -1,2 +1,3 @@
 mod charge_gas;
+mod json_schema;
 mod metered_utils;